requirements.txt
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.7
imageio==2.34.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.3
PyMatting==1.1.12
referencing==0.35.1
rembg==2.0.58
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.0
sympy==1.13.2
tifffile==2024.8.10
tqdm==4.66.5
urllib3==2.2.2
